Move lib/juice.js to index.js
Removes unnecessary indirection of index.js.
